Integration of Feedback during Certification
Angekreuzt and Klient: changed strings for responses.
Klient:
- response for no answer from politician now includes party of politician and reference to abgeordnetenwatch.de
Angekreuzt: - empty slot values, which create strings only containing a null pointer are now handled

classes ArbeitundSoziales, SteuernundFinanzen, Bildung, EUundAussenpolitik,IntegrationundAsyl, Sicherheit : method extractInformation extracted and removed to Delegate.extractInformation

class Zweitstimme now also uses the mapping from Categorys to Groups

all classes: subGroup renamed to Category, which is according to the Json file from abgeordnetenwatch.de.Categorys are included in Groups

+++ src/SpeechHelper.java
@@ -17,7 +17,7 @@ public class SpeechHelper {
  {return "<say-as interpret-as=\""+type+"\">"+s+"</say-as>";}
 
 
- public static String createBreak(int seconds)
+ public static String createBreak(double seconds)
  {return "<break time=\""+seconds+"s\"/>";	}
  
  }
